Crossterm is working in the new environment.
