+++ bracket-terminal/Cargo.toml
@@ -36,7 +36,7 @@ winit = { version = "0.21.0", optional = true }
 glutin = {version = "0.23.0", optional = true }
 
 [features]
-default = [ "curses" ]
+default = [ "crossterm" ]
 opengl = [ "glow", "image", "glutin" ]
 curses = [ "pancurses", "winit" ]
 amethyst_engine_vulkan = [ "amethyst/vulkan", "image" ]
