+++ bracket-terminal/src/hal/crossterm_be/shader.rs
@@ -1,0 +1,1 @@
+pub struct Shader {}
